fix time overflow bug in tidy-finger-scroll
button_release_event_cb wants to know the speed at which the scrolling surface was traveling
on release by the users finger to allow it to continue scrolling at this speed until it
coasts to a stop depending on the deceleration strength. To avoid noise in the output of
the ts it takes the mean speed of the last 3 motion events. To calculate this mean speed it
calculates the mean position and start time of the events. To calculate the mean time it adds
the time.tv_sec of the motion events together and divides by 3. Since Fri May 15 02:09:25
2015 UTC this causes tv_sec to overflow with the final overflowed result of the calculation
being that the motion event started somewhere close to 1901. The effect of this is that the
scrolling surface stops immediately after released by the finger, most notably this happens
when scrolling in the launcher, but other places like tasknav are allso affected.
This pr fixes the problem by replacing the accumulating variable tv_sec with a 64bit quantity