hd-launcher-editor: fix reordering for Main menu
Ever since commit 604d1167860d5750fffe097de121bd7a3e2885f7 ("Implement
apt hook for generating a Debian xdg menu."), icon reordering was
broken.

The problem is that the hildon-desktop's launcher editor does not seem
to have been written with support for nested menus as introduced in this
commit: it will just override $HOME/.config/menus/hildon.menu with
whatever icon order is available in the current context.

The code in this commit will reject any icons that aren't in the "Main"
category (/usr/share/applications/hildon), which mostly means it will
exclude everything in the "Debian" category. As a result, editing icons
in the main menu will work as expected.

In the future, we might want to make the launcher editor aware of
submenus and significantly rewrite the editing of the menu files - since
the editor would need full context to not stomp on previous user
changes. Perhaps it would be possible to write a
$HOME/.config/menus/debian.menu file and include that from the
$HOME/.config/menus/hildon.menu file, using the MergeFile instruction.
But it would still be limited to one submenu, the Debian one.