hd-laucher: disable editing non-top pages
This is a follow-up commit of:

    "hd-launcher-editor: fix reordering for Main menu"

to disallow editing in the non-top page, since this is not currently
supported by the code. However, since commit:

    "hildon-menu-generate: create separate debian.menu"

we could edit the hd-launcher-editor code to be able to write a
separate $HOME/.config/menus/debian.menu file (or really, any file given
the category it is working on) to support editing sub-menus.