hd-launcher-editor: support Debian submenu editing
Introduce a new "category" property in hd-launcher-editor, and use that
category to write sub menu entries.

The code is general enough to technically support other submenus beyond
Debian, but to actually get those to show properly, we still need to add
them to the main hildon.menu file in the user XDG dir, and I'm not quite
sure what that would look like, so let's leave that for later.